+++ salesforce/soap_bodies.py
@@ -216,7 +216,7 @@ retrieve_sobjects_workflow_task_body = """
   </soapenv:Envelope>"""
 
 # Body for retrieving static resource
-retrieve_static_resources_body = """
+retrieve_static_resource_body = """
     <soapenv:Envelope xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/" 
     xmlns:met="http://soap.sforce.com/2006/04/metadata">
         <soapenv:Header>
@@ -232,6 +232,47 @@ retrieve_static_resources_body = """
                         <met:types>
                             <met:members>*</met:members>
                             <name>StaticResource</name>
+                        </met:types>
+                        <met:version>{1}.0</met:version>
+                    </met:unpackaged>
+                </met:retrieveRequest>
+            </met:retrieve>
+        </soapenv:Body>
+    </soapenv:Envelope>"""
+
+# Body for retrieving static resource
+retrieve_apex_code_body = """
+    <soapenv:Envelope xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/" 
+    xmlns:met="http://soap.sforce.com/2006/04/metadata">
+        <soapenv:Header>
+            <met:SessionHeader>
+               <met:sessionId>{0}</met:sessionId>
+            </met:SessionHeader>
+        </soapenv:Header>
+        <soapenv:Body>
+            <met:retrieve>
+                <met:retrieveRequest>
+                    <met:apiVersion>{1}.0</met:apiVersion>
+                    <met:unpackaged>
+                        <met:types>
+                            <met:members>*</met:members>
+                            <name>StaticResource</name>
+                        </met:types>
+                        <met:types>
+                            <met:members>*</met:members>
+                            <name>ApexClass</name>
+                        </met:types>
+                        <met:types>
+                            <met:members>*</met:members>
+                            <name>ApexPage</name>
+                        </met:types>
+                        <met:types>
+                            <met:members>*</met:members>
+                            <name>ApexTrigger</name>
+                        </met:types>
+                        <met:types>
+                            <met:members>*</met:members>
+                            <name>ApexComponent</name>
                         </met:types>
                         <met:version>{1}.0</met:version>
                     </met:unpackaged>